requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-jazzmin==2.5.0
djangorestframework==3.14.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
python-dateutil==1.5
python-decouple==3.6
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
whitenoise==6.2.0
